refactor(providers): 将系统提供商配置移动到config/providers文件
